Extract SprintStatistics class to its own file
